refactor : directory 위치 변경 및 불필요한 file 삭제
