--- backend/src/main/java/com/techeer/backend/feedback/Repository/ResumeRepository.java
@@ -1,8 +1,0 @@
-package com.techeer.backend.feedback.Repository;
-
-
-import com.techeer.backend.feedback.entity.Resume;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ResumeRepository extends JpaRepository<Resume, Long> {
-}
